combine delete docs test files
